requirements.txt
streamlit>=1.20.0
requests>=2.28.0
speechrecognition>=3.8.1
pydub>=0.25.1
python-dotenv>=1.0.0
